refactor(evidence): Standardize file structure and naming conventions
- Rename Components directory to lowercase components
- Update import paths to reflect new directory structure
- Reorganize actions and hooks imports
- Migrate edit evidence page to use client component with server-side data fetching
- Apply consistent formatting and trailing comma rules
