+++ apps/app/.vscode/settings.json
@@ -1,0 +1,9 @@
+{
+    "editor.defaultFormatter": "biomejs.biome",
+    "editor.codeActionsOnSave": {
+          "quickfix.biome": "explicit"
+      },
+      "recommendations": [
+		"biomejs.biome"
+	]
+  }
