+++ apps/app/src/app/[locale]/(app)/(dashboard)/evidence/[id]/Components/EvidenceDetails.tsx
@@ -1,12 +1,8 @@
 "use client";
 
 import { Alert, AlertDescription, AlertTitle } from "@bubba/ui/alert";
-import { Button } from "@bubba/ui/button";
 import { Skeleton } from "@bubba/ui/skeleton";
 import { CheckCircle2, FileIcon, XCircle } from "lucide-react";
-import { useAction } from "next-safe-action/hooks";
-import { toast } from "sonner";
-import { publishEvidence } from "../Actions/publishEvidence";
 import { useOrganizationEvidence } from "../hooks/useOrganizationEvidence";
 import type { EvidenceDetailsProps } from "../types";
 import { ReviewSection } from "./ReviewSection";
