+++ apps/app/src/app/[locale]/(app)/(dashboard)/evidence/[id]/Components/ReviewSection.tsx
@@ -8,8 +8,8 @@ import { format } from "date-fns";
 import { Building, CalendarClock, RefreshCw, User } from "lucide-react";
 import { useAction } from "next-safe-action/hooks";
 import { toast } from "sonner";
-import { publishEvidence } from "../Actions/publishEvidence";
-import { toggleRelevance } from "../Actions/toggleRelevance";
+import { publishEvidence } from "../actions/publishEvidence";
+import { toggleRelevance } from "../actions/toggleRelevance";
 import { useOrganizationEvidence } from "../hooks/useOrganizationEvidence";
 import { AssigneeSection } from "./AssigneeSection";
 import { DepartmentSection } from "./DepartmentSection";
