+++ apps/app/src/app/[locale]/(app)/(dashboard)/evidence/[id]/hooks/useUrlManagement.ts
@@ -2,7 +2,7 @@
 
 import { useState, useCallback } from "react";
 import { useToast } from "@bubba/ui/use-toast";
-import { updateEvidenceUrls } from "../Actions/updateEvidenceUrls";
+import { updateEvidenceUrls } from "../actions/updateEvidenceUrls";
 import type { ActionResponse } from "@/actions/types";
 
 interface DraftUrl {
@@ -62,7 +62,7 @@ export function useUrlManagement({
 
   const handleUpdateDraft = useCallback((id: string, url: string) => {
     setDraftUrls((prev) =>
-      prev.map((draft) => (draft.id === id ? { ...draft, url } : draft)),
+      prev.map((draft) => (draft.id === id ? { ...draft, url } : draft))
     );
   }, []);
 
@@ -151,7 +151,7 @@ export function useUrlManagement({
         });
       }
     },
-    [currentUrls, evidenceId, onSuccess, toast],
+    [currentUrls, evidenceId, onSuccess, toast]
   );
 
   return {
