+++ apps/app/src/app/[locale]/(app)/(dashboard)/evidence/[id]/page.tsx
@@ -1,4 +1,4 @@
-import { EvidenceDetails } from "./Components/EvidenceDetails";
+import { EvidenceDetails } from "./components/EvidenceDetails";
 
 interface EvidencePageProps {
   params: Promise<{
